+++ src/util/algs.hh
@@ -14,7 +14,7 @@
 namespace COMPARE {
 short compare(clong& x, clong& y);
 
-short compare_map(const map<ushort, ushort>& m1, const map<ushort, ushort>& m2);
+short compare_map(const map<uint, ushort>& m1, const map<uint, ushort>& m2);
 template<class T>
 short compare_container(const T& x, const T& y);
 }
